tests: refactor duplicated code
