+++ tests/unit_tests/conftest.py
@@ -1,6 +1,5 @@
 import pytest
 from aioresponses import aioresponses
-from async_substrate_interface.sync_substrate import SubstrateInterface
 
 import bittensor.core.subtensor
 
@@ -17,18 +16,17 @@ def mock_aio_response():
 
 
 @pytest.fixture
-def mock_substrate_interface(mocker):
-    mocked = mocker.MagicMock(
-        autospec=SubstrateInterface,
+def mock_substrate(mocker):
+    mocked = mocker.patch(
+        "bittensor.core.subtensor.SubstrateInterface",
+        autospec=True,
     )
 
-    mocker.patch("bittensor.core.subtensor.SubstrateInterface", return_value=mocked)
-
-    return mocked
+    return mocked.return_value
 
 
 @pytest.fixture
-def subtensor(mock_substrate_interface):
+def subtensor(mock_substrate):
     return bittensor.core.subtensor.Subtensor()
