+++ tests/unit_tests/test_subnets.py
@@ -1,5 +1,4 @@
 import pytest
-from mpmath.ctx_mp_python import return_mpc
 
 from bittensor.utils import subnets
